Update expect action for clarity
Summary:
The console output displayed when using `expect` action was cumbersome and a little difficult to interpret.

Expect function by itself, is good enough to display the input and output succintly.
It seemed like having additional logging made it redundant and unclear to interpret.

Cleaning up the logging so that output from `expect` action is easy to read.

Differential Revision: D60419102