Update expect action for clarity
Summary:
The console output displayed when using `expect` action was cumbersome and a little difficult to interpret.

Expect function by itself, is good enough to display the input and output succintly.

Changing the logging level to DEBUG, so that it can be displayed if required when calling the `expect` action

Differential Revision: D60423533