Update expect action for clarity (#505)
Summary:
Pull Request resolved: https://github.com/facebookincubator/TTPForge/pull/505

The console output displayed when using `expect` action was cumbersome and a little difficult to interpret.

Expect function by itself, is good enough to display the input and output succintly.

Changing the logging level to DEBUG, so that it can be displayed if required when calling the `expect` action

Reviewed By: nicolagiacchetta

Differential Revision: D60423533

fbshipit-source-id: 5dad159b44cd4796ea593861484b1baad1595565