Create mutable variable store (#514)
Summary:
Pull Request resolved: https://github.com/facebookincubator/TTPForge/pull/514

Create a variable store that we can mutate between test steps.  Move WorkDir into this variable store so that we can mutate the work dir for a future cd step.

In the future if we want to carry any other information between steps in a mutable fashion, all we need to do is add a new variable to this struct.

Reviewed By: inesusvet

Differential Revision: D63457210