Validate CD Step at Run instead of Validate (#519)
Summary:
Pull Request resolved: https://github.com/facebookincubator/TTPForge/pull/519

Because TTPForge runs validations on TTP load, and we validate our target directory in the Validate action, the CD step will validate the existence of its target directory at TTP load.  This means that if you create your target directory as part of your TTP, your CD step will fail because the directory does not exist yet.

By moving the check to the run step, we will now hold off on validating the directory's existence until the CD step is run, allowing previous steps to create the target directory.

Reviewed By: inesusvet

Differential Revision: D65831540

fbshipit-source-id: dfbcb0e6f14ab93ad2865691fb0b7c0d48beba38