Enum Command (#543)
Summary:
Pull Request resolved: https://github.com/facebookincubator/TTPForge/pull/543

### TTPForge Enum Command

#### Overview

This diff introduces a new `enum` command to the TTPForge framework. It is designed to provide a way to enumerate TTPs, making it easier for users to discover and manage TTPs. With this implementation, users can display platform wise TTP count across repositories with filtering on platforms, tactics, techniques and sub-techniques with an option to view specific TTPs with a verbose output.

#### Context

There was a growing need to effectively enumerate, dynamically filter and count TTPs across repositories created for various purple team engagements. This diff solves aspects of this problem paving way for further advancements. This capability enables users to maintain tally of TTPs across engagements and repositories to allow for better management ultimately leading to improved decision-making and resource allocation.

#### Future Improvements

The addition of TTP dynamic enumeration and counting capabilities in TTPForge paves for the following future advancements:

1) Unified View: The ability to enumerate and count TTPs paves the way for creating an interactive UI that displays all TTPs from multiple repositories in one place. This provides managers and stakeholders with a holistic view of their TTP landscape with super helpful filtering mechanisms, empowering them to make data-driven decisions and allocate resources more effectively.

2) Reducing redundancy: By viewing all TTPs in one place with advanced filtering would help reduce redundancy and increase efficiency of developers searching for similar TTPs.

3) Analysis: The addition of TTP enumeration and counting capabilities sets the stage for future enhancements, such as advanced filtering, sorting, and analytics, which will further empower our team to pinpoint how our TTP catalog could improve further.

4) Enhanching enumeration: Adding new enum subcommands to enumerate other things like repo.

5) Changing from regex to parsing to improve enumeration & tagging: [Document](https://docs.google.com/document/d/1OUihSxvrTHUK24kIH0VE3uHL-l6OBCL6iSPLIXLtx8E/edit?tab=t.0)

By addressing the gap in TTPForge commands, this update significantly enhances the tool's functionality, and overall value proposition, ultimately benefiting users and organizations that rely on TTPs for various purposes.

#### Changes

*   The `enum` command is added to the root command.
*   `enum.go` and `enum_test.go` hold the core logic of implementation and unit tests.
*   New test repository `enum-repo` is created with sample TTPs with test configuration in `test-config.yaml` for unit testing.

#### References
1) https://pkg.go.dev/github.com/spf13/cobra#Command
2) https://docs.google.com/document/d/1YtnMDvg7uEhWuEMmSNpvs0JdQbAmXK3f1g81Ortso2g/edit?tab=t.0#heading=h.qzjpd1hy8nzl

Reviewed By: d0n601, andreizharinov

Differential Revision: D77619765

fbshipit-source-id: 74c540b79d6c36c78ea42afe02b4707b18a441e2