TTPForge StepVars Bugfix: Fix newline appended to stdout
Summary:
In inline blocks, when we capture stdout to a variable, we include the newline character used to move the cursor to a fresh line.  While this is technically correct, most users would expect something like `echo "asdf"` to load the variable with `asdf`, not `asdf\n`.

To fix this, I've opted just to remove a single newline from the end of stdout before loading it into the variable.

Differential Revision: D78497399