TTPForge StepVars Bugfix: Fix newline appended to stdout (#544)
Summary:
Pull Request resolved: https://github.com/facebookincubator/TTPForge/pull/544

In inline blocks, when we capture stdout to a variable, we include the newline character used to move the cursor to a fresh line.  While this is technically correct, most users would expect something like `echo "asdf"` to load the variable with `asdf`, not `asdf\n`.

To fix this, I've opted just to remove a single newline from the end of stdout before loading it into the variable.

Reviewed By: d0n601

Differential Revision: D78497399

fbshipit-source-id: 41ab4c783f2d6ec0ec8d7754a58c6feb4f1f7c7e