Adding parse-yaml functionality for sandcastle job
Summary:
### TTPForge Enhancement: Parse-YAML Functionality for Sandcastle Job

This diff introduces a new command `parse-yaml` to the TTPForge tool. The command accepts TTP reference as an argument and reads & tries to parse them using the `parseutils`. In case of any failure it returns an error. In this diff we also create a bash file that is called by the sandcastle job with all the affected file names. The bash files calls parse-yaml functionality for the TTPs in a for loop. This enables the sandcastle job to check & ensure that all the affected TTPs have correct structure.

#### New Files Added

* `fbcode/security/redteam/purple_team/ttpforge/cmd/parseyaml.go`: Implements the parse-YAML functionality.
* `fbcode/security/redteam/purple_team/ttpforge/cmd/parseyaml_test.go`: Contains unit tests for the parse-YAML feature.
* `fbcode/security/redteam/purple_team/ttpforge/parse-diff-ttps-test.sh`: Script called by sandcastle job that runs the parse-YAML functionality for all TTPs given to it.

These changes collectively enhance the functionality of TTPForge by adding a command to parse YAML files and making it more versatile for Sandcastle jobs.

Reviewed By: ivnik

Differential Revision: D79202811