TTPForge Move Command Reference Update (#553)
Summary:
Pull Request resolved: https://github.com/facebookincubator/TTPForge/pull/553

Updated TTPForge Move command to consider the destination repository when performing the move.

Now updates files to reference `//path/to/ttp.yaml` for the new subTTP path instead of `repo//path/to/ttp.yaml` by default.

If the destination repo is different than the repo the file is in, a repository name will be provided based off the one in the config. **This has the potential to create inconsistencies in name resolution as different users may have different mappings from the standard ones in the config file.**

Updated documentation to discuss this change and provide a warning for cross-repository moves.

Reviewed By: samdlinn

Differential Revision: D81613044

fbshipit-source-id: aa64eae9d188be8de1bafb761eaa1305da808497