TTPForge Enum Dependencies Command
Summary:
# Overview
This diff introduces an addition to the `enum` command `enum dependencies` that will enumerate all TTPs that use a given TTP as a subTTP step.

# Context
This implementation spawned from a need to identify dependencies within the growing list of available forges.

# Future Improvements
The testing suite sets up several possible scenarios, but it could be improved by confirming command output matches what's expected. This should be made easier if all modules migrate to the internal logging package.

# Changes
- Added `dependencies` subcommand to the `enum` command.
- `enumdependencies.go` and `enumdependencies_test.go` hold implementation and testing for the `enum dependencies` command respectively.

Reviewed By: RoboticPrism

Differential Revision: D82462426