TTPForge Remove TTP Command (#557)
Summary:
Pull Request resolved: https://github.com/facebookincubator/TTPForge/pull/557

# Overview
This diff introduces an addition to the `remove` command `remove ttp` that will safely remove a TTP file by checking for dependencies before removal.

# Context
This implementation spawned from a need to identify dependencies when managing files within the growing list of available forges.

# Future Improvements
A basic test suite could be added to confirm that files are being removed appropriately with the expected output. This will be made easier with the internal logging package.

# Changes
- Added `ttp` subcommand to the `remove` command in `removettp.go`.
- Added `remove.md` documenting the remove command and its subcommands.

Reviewed By: RoboticPrism

Differential Revision: D82462428

fbshipit-source-id: f922176cd28ff3825504c578efaa05eeb553c43d