TTPForge: Enable StepVars in Own Cleanup Action (#560)
Summary:
Pull Request resolved: https://github.com/facebookincubator/TTPForge/pull/560

Currently, StepVars can only be used in actions after the action they are created in.  This is also true of cleanup actions, meaning that you cannot use a StepVar generated in an action in that actions own cleanup.

By by waiting to do templating until we're at cleanup time, we can allow a StepVar created in a step to be available within its own cleanup action.

An example of where this could be useful could be a TTP that calls an API to create a new user and recieves a dynamically created new user id.  Now we can get the user id in the action, and then pass the  user id into the cleanup step to call a user deletion API.

Reviewed By: d0n601

Differential Revision: D83680470

fbshipit-source-id: 7baa5e3c3a08815b8a7f6e651a26b43f590cfac9