Bugfix: Logger config not being respected
Summary:
We currently make two calls to initialize the logger in this order:
- when the logger package is imported with a default configuration
- when we have initialized the configuration with the config we just initialized

However, currently we have wrapped the logger initialization in a do once.  This led to us always setting up a logger with default configurations and ignoring the second call with the configs we actually want.

I opted to remove the init function (which is called when the package is loaded) and instead have the call for the global logger create a new empty config logger if one does not exist yet.   In this case, this means we'll always yield to a provided config, but also that tests can just request and receive a standard logger without needing to create one manually in every test.

Differential Revision: D84361195