HTTPRequest Block Response Headers (#564)
Summary:
### What?
This diff adds a feature to the HTTPRequest block which allows a user to specify whether or not they'd like the server headers returned as part of the response. The option defaults to `false` so as not to break existing TTPs which are built not to expect headers as part of the response.

### Why?
Retrieving headers is a critical part of many attack flows, particularly ones that involve gathering cookies to manage sessions. It is also a requirement when a user makes a `HEAD` type request.

Differential Revision: D84760010

Pulled By: d0n601

fbshipit-source-id: 8a8abb0b8a903c393baa426d285fcb4b6bbd3fda