Fix bug with path arg type and add env var expansion (#571)
Summary:
Pull Request resolved: https://github.com/facebookincubator/TTPForge/pull/571

The path arg type was not being evaluated based on the path of the TTP YAML. This caused relative paths such as `default: <FILE>` or `default: ./<FILE>` to resolve based on the path you executed ttpforge from as opposed to the path of the YAML.

This has been updated to resolve relative to the YAML when provided as a default value and resolve relative to the current working directory when provided as a command-line argument.

Additionally added handling for variable expansion within the shell. This should be helpful for TTPs referencing variables like $HOME or alike.

Reviewed By: RoboticPrism

Differential Revision: D85156301

fbshipit-source-id: 69d8db00c37576684a03ef30d0d55ba078ec2e61